Unify story data caching in data_io
